requirements.txt
vosk>=0.3.45
faster-whisper>=1.1.0
openai-whisper>=20231117
sounddevice>=0.5.0
soundcard>=0.4.0
numpy>=1.26.0
pynput>=1.8.0
argostranslate>=1.9.0
openai>=1.0.0
anthropic>=0.40.0
deepgram-sdk>=6.0.0
customtkinter>=5.2.0
